+++ benchmark/utils/time_from_image_name.py
